+++ app/Http/Controllers/ProductController.php
@@ -55,6 +55,7 @@ class ProductController extends Controller
     public function manage(Product $product)
     {
         $collection = $product->orderBy('created_at', 'desc')->paginate(10);
+        $this->authorize($collection);
         return view('backoffice.productDashboard.manage', ['collection' => $collection]);
     }
 
@@ -68,6 +69,7 @@ class ProductController extends Controller
     {
         $request->validate($this->rules_delete);
         $product = $product->find((int)$request->id);
+        $this->authorize($product);
         if(isset($product->id))
             $product->delete(); 
         return redirect(route('manageProduct'));
@@ -82,9 +84,9 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-
         $this->validate($request, $this->rules);
         $product = new Product;
+        $this->authorize($product);
         $product->user_id = Auth::user()->id;
         $product->name = $request->title;
         $product->description = ($request->description);
@@ -112,6 +114,7 @@ class ProductController extends Controller
      */
     public function updateView(Request $request,Product $product)
     {
+        $this->authorize('updateView',$product);
         $request->validate($this->rules_delete);
         $item = $product::find($request->id);
         return view('backoffice.productDashboard.update', ['item' => $item]);
@@ -126,6 +129,7 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+        $this->authorize($product);
         $request->validate($this->rules_update, $this->errorMessages_update);
         $product = $product->find((int)$request->id);
         $product->exists=true;
